support inverted observations for gRPC
