+++ packages/streams-adapter/redcon/redcon.go
@@ -243,9 +243,21 @@ func (s *RedconServer) handleEval(conn redcon.Conn, cmd redcon.Command) {
 	if s.publisher != nil {
 		if rawKeys, ok := s.cache.RawKeysByTransformed(transformedKey); ok {
 			for _, rawKey := range rawKeys {
-				if payloadHash, ok := s.cache.PayloadHashByRawKey(rawKey); ok {
-					s.publisher.Publish(payloadHash, obs, ts)
+				payloadHash, ok := s.cache.PayloadHashByRawKey(rawKey)
+				if !ok {
+					continue
 				}
+				out := obs
+				if item := s.cache.Get(rawKey); item != nil && item.RequiresInverse && obs.Success {
+					inverted, err := helpers.InvertObservation(obs)
+					if err != nil {
+						s.logger.Error("failed to invert observation for inverse subscriber",
+							"rawKey", rawKey, "error", err)
+						continue // do not publish a direct price to an inverse subscriber
+					}
+					out = inverted
+				}
+				s.publisher.Publish(payloadHash, out, ts)
 			}
 		}
 	}
